requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.34.0
jinja2==3.1.3
pydantic==2.11.5
python-dotenv==1.0.1
langchain==0.1.16
langchain-community==0.0.21
langchain-core==0.1.37
langchain-groq==0.0.6
langchain-huggingface==0.0.4
sentence-transformers==2.6.1
transformers==4.40.0
faiss-cpu==1.7.4
pypdf==3.17.4


# pip install -r requirements.txt